+++ src/domible/builders/__init__.py
@@ -1,9 +1,10 @@
-""" domible/src/domible/builders/__init__.py """
+""" domible/builders/__init__.py """
 
-from domible.builders.tableBuilder import RowBuilder, TableBuilder, build_table_from_dicts
-from domible.builders.navBuilder import NavBuilder 
-from domible.builders.elementFromObject import element_from_object  
-from domible.builders.formBuilder import (
+from .tableBuilder import RowBuilder, TableBuilder, build_table_from_dicts
+from .navBuilder import NavBuilder 
+from .elementFromObject import element_from_object  
+from .preformatted import python_code_block, python_code_style 
+from .formBuilder import (
     ToggleAllDetailsButton,
     ExpandAllDetailsButton,
     CollapseAllDetailsButton,