+++ src/domible/elements/__init__.py
@@ -1,19 +1,19 @@
-""" domible/src/domible/elements/__init__.py
+""" domible/elements/__init__.py
 """
 
 ##
 # only expose base classes for typing 
 # for example, cases where a container can operate on any BaseClass derived element.
 ##
-from domible.elements.baseElements import BaseElement, BaseVoidElement 
+from .baseElements import BaseElement, BaseVoidElement 
 
-from domible.elements.roots import Html
-from domible.elements.meta import Base, Head, Meta, Style, Title 
-from domible.elements.inline import Anchor, Div, Paragraph, Span, Summary, Details
-from domible.elements.lists import DescriptionTerm, DescriptionDef, DListItem, ListItem, OrderedList, UnorderedList, MenuList, DescriptionList
-from domible.elements.scripting import Canvas, NoScript, Script, Template 
-from domible.elements.sectioning import Article, Body, Footer, Header, Heading, HorizontalRule, Main, Nav, Section 
-from domible.elements.tables import TableData, TableHeader, TableColumnHeader, TableRowHeader, TableRow, TableHead, TableBody, TableFoot, Caption, Table
-from domible.elements.forms import Button 
+from .roots import Html
+from .meta import Base, Head, Meta, Style, Title 
+from .inline import Anchor, Code, Div, Paragraph, Pre, Span, Summary, Details
+from .lists import DescriptionTerm, DescriptionDef, DListItem, ListItem, OrderedList, UnorderedList, MenuList, DescriptionList
+from .scripting import Canvas, NoScript, Script, Template 
+from .sectioning import Article, Body, Footer, Header, Heading, HorizontalRule, Main, Nav, Section 
+from .tables import TableData, TableHeader, TableColumnHeader, TableRowHeader, TableRow, TableHead, TableBody, TableFoot, Caption, Table
+from .forms import Button 
 
 ## end of file