+++ src/domible/scripts/__init__.py
@@ -1,0 +1,1 @@
+# domible/scripts/__init__.py