+++ src/domible/scripts/dibrowse/__init__.py
@@ -1,0 +1,1 @@
+# domible/scripts/dibrowse/__init__.py