+++ src/domible/scripts/dicli/__init__.py
@@ -1,0 +1,1 @@
+# domible/scripts/dicli/__init__.py