+++ src/domible/starterDocuments/__init__.py
@@ -1,5 +1,5 @@
-""" domible/src/domible/starterDocuments/__init__.py """
+""" domible/starterDocuments/__init__.py """
 
-from domible.starterDocuments.basicDocuments import basic_head_empty_body
+from .basicDocuments import basic_head_empty_body
 
 # end of file